+++ tests/testthat/test-2-get_tiles_3dep.R
@@ -1,12 +1,15 @@
 test_that("get_tiles gets the same elevation tiles twice", {
   skip_on_cran()
-  output_tif <- get_tiles(list(
-    c(lat = 44.04905, lng = -74.01188),
-    c(lat = 44.04911, lng = -74.01179)
-  ),
-  services = c("elevation", "3DEPElevation"),
-  georeference = FALSE
+  dl_loc <- data.frame(
+    lat = c(44.04905, 44.04911),
+    lng = c(-74.01188, -74.01179)
   )
+  dl_loc <- sf::st_as_sf(dl_loc, coords = c("lng", "lat"))
+  sf::st_crs(dl_loc) <- sf::st_crs(4326)
+
+  output_tif <- get_tiles(dl_loc,
+                          services = c("elevation", "3DEPElevation"),
+                          georeference = FALSE)
 
   expect_equal(names(output_tif), "elevation")
 
@@ -25,25 +28,23 @@ test_that("get_tiles gets the same elevation tiles twice", {
 })
 
 test_that("get_tiles fails as expected", {
-  expect_error(get_tiles(list(
-    c(lat = 44.04905, lng = -74.01188),
-    c(lat = 44.04911, lng = -74.01179)
-  ),
+  dl_loc <- data.frame(
+    lat = c(44.04905, 44.04911),
+    lng = c(-74.01188, -74.01179)
+  )
+  dl_loc <- sf::st_as_sf(dl_loc, coords = c("lng", "lat"))
+  sf::st_crs(dl_loc) <- sf::st_crs(4326)
+
+  expect_error(get_tiles(dl_loc,
   services = c("elevation", "ortho", "USGSNAIPPlus"),
   side_length = 4097
   ))
 
-  expect_error(get_tiles(list(
-    c(lat = 44.04905, lng = -74.01188),
-    c(lat = 44.04911, lng = -74.01179)
-  ),
+  expect_error(get_tiles(dl_loc,
   services = c("exactly_the_data_i_want")
   ))
 
-  expect_error(get_tiles(list(
-    c(lat = 44.04905, lng = -74.01188),
-    c(lat = 44.04911, lng = -74.01179)
-  ),
+  expect_error(get_tiles(dl_loc,
   services = c("elevation"),
   side_length = 8001
   ))