+++ tests/testthat/test-2-get_tiles_contours.R
@@ -1,9 +1,13 @@
 test_that("get_tiles gets the same contours tiles twice", {
   skip_on_cran()
-  output_tif <- get_tiles(list(
-    c(lat = 44.04905, lng = -74.01188),
-    c(lat = 44.04911, lng = -74.01179)
-  ),
+  dl_loc <- data.frame(
+    lat = c(44.04905, 44.04911),
+    lng = c(-74.01188, -74.01179)
+  )
+  dl_loc <- sf::st_as_sf(dl_loc, coords = c("lng", "lat"))
+  sf::st_crs(dl_loc) <- sf::st_crs(4326)
+
+  output_tif <- get_tiles(dl_loc,
   services = c("contours"),
   georeference = FALSE
   )